% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcSampleCov}
\alias{calcSampleCov}
\title{Calculate sample covariance}
\usage{
calcSampleCov(data, sample_mean, N, P)
}
\arguments{
\item{data}{Data in matrix format}

\item{sample_mean}{Sample mean for data}

\item{n}{The number of samples in data}

\item{n_col}{The number of columns in data}
}
\value{
One of the parameters required to calculate the posterior of the
 Multivariate normal with unknown mean and covariance (the unnormalised
 sample covariance).
}
\description{
Returns the unnormalised sample covariance. Required as
arma::cov() does not work for singletons.
}
