% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMDI.R
\name{callMDI}
\alias{callMDI}
\title{Call MDI}
\usage{
callMDI(
  X,
  R,
  thin,
  initial_labels,
  fixed,
  types,
  K,
  alpha = NULL,
  initial_labels_as_intended = FALSE
)
}
\arguments{
\item{X}{Data to cluster. List of matrices with the N items to cluster held
in rows.}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. if
``thin=50`` only every 50th sample is kept.}

\item{initial_labels}{Initial clustering. $N x L$ matrix.}

\item{fixed}{Which items are fixed in their initial label. $N x L$ matrix.}

\item{alpha}{The concentration parameter for the stick-breaking prior and the
weights in the model.}

\item{initial_labels_as_intended}{Logical indicating if the passed initial 
labels are as intended or should ``generateInitialLabels`` be called.}

\item{type}{Character vector indicating density type to use. 'MVN'
(multivariate normal), 'TAGM' (t-adjust Gaussian mixture) or 'C' (categorical).}

\item{K_max}{Vector indicating the number of components to include (the upper
bound on the number of clusters in each dataset).}
}
\value{
A named list containing the sampled partitions, component weights,
phi and mass parameters, model fit measures and some details on the model call.
}
\description{
Runs a MCMC chain of the integrative clustering method,
Multiple Dataset Integration (MDI), to L datasets. It is recommended that
L < 5.
}
\examples{

N <- 100
X <- matrix(c(rnorm(100, 0, 1), rnorm(100, 3, 1)), ncol = 2)
Y <- matrix(c(rnorm(100, 0, 1), rnorm(100, 3, 1)), ncol = 2)
data_modelled <- list(X, Y)

R <- 100
thin <- 5

alpha <- c(1, 1)
K <- c(10, 15)
labels <- matrix(0, nrow = N, ncol = 2)
labels[, 1] <- generateInitialUnsupervisedLabels(N, alpha[1], K[1])
labels[, 2] <- generateInitialUnsupervisedLabels(N, alpha[1], K[2])

fixed <- matrix(0, nrow = N, ncol = 2)
types <- c("MVN", "G")

mcmc_out <- callMDI(data_modelled, R, thin, labels, fixed, types, K, alpha)

}
