% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callMixtureModel.R
\name{callMixtureModel}
\alias{callMixtureModel}
\title{Call Mixture Model}
\usage{
callMixtureModel(
  X,
  R,
  thin,
  type,
  K = 75,
  initial_labels = NULL,
  fixed = NULL,
  alpha = NULL,
  initial_labels_as_intended = FALSE
)
}
\arguments{
\item{X}{Data to cluster. Matrix with the N items to cluster held
in rows.}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. if
``thin=50`` only every 50th sample is kept.}

\item{type}{Character vector indicating density type to use. 'MVN'
(multivariate normal), 'TAGM' (t-adjust Gaussian mixture) or 'C' (categorical).}

\item{initial_labels}{Initial clustering. $N$-vector.}

\item{fixed}{Which items are fixed in their initial label. $N$-vector.}

\item{alpha}{The concentration parameter for the stick-breaking prior and the
weights in the model.}

\item{initial_labels_as_intended}{Logical indicating if the passed initial
labels are as intended or should ``generateInitialLabels`` be called.}

\item{K_max}{Integer indicating the number of components to include (the upper
bound on the number of clusters).}
}
\value{
A named list containing the sampled partitions, component weights,
phi and mass parameters, model fit measures and some details on the model call.
}
\description{
Runs a MCMC chain of a Bayesian mixture model. Essentially a
wrapper to allow more intuitive inputs for the single dataset case of MDI.
}
\examples{
N <- 100
X <- matrix(c(rnorm(100, 0, 1), rnorm(100, 3, 1)), ncol = 2)

R <- 100
thin <- 5

alpha <- 1
K <- 10
type <- "MVN"

mcmc_out <- callMixtureModel(X, R, thin, type, K = K)

}
