% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataCorrectInput.R
\name{checkDataCorrectInput}
\alias{checkDataCorrectInput}
\title{Check data correct input}
\usage{
checkDataCorrectInput(X)
}
\arguments{
\item{X}{Data passed to ``callMDI``. Should be a list of matrices each with 
N items held in rows.}
}
\description{
Internal function that checks the data passed to ``callMDI`` is
the correct format.
}
