% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFixedInput.R
\name{checkFixedInput}
\alias{checkFixedInput}
\title{Check fixed input}
\usage{
checkFixedInput(fixed, N, V)
}
\arguments{
\item{fixed}{$N x V$ matrix indicating which labels are observed and hence 
fixed. If the vth column has no 1's then this view is unsupervised.}

\item{N}{The number of items being clustered.}

\item{V}{The number of views being modelled.}
}
\value{
No return value, called for side effects.
}
\description{
Checks if the ``fixed`` input for ``callMDI`` has the expected 
properties.
}
\examples{
N <- 100
V <- 3
fixed <- matrix(0, N, V)
checkFixedInput(fixed, N, V)

}
