% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLabels.R
\name{checkLabels}
\alias{checkLabels}
\title{Check labels}
\usage{
checkLabels(labels, K)
}
\arguments{
\item{labels}{Vector of initial labels for a view in ``callMDI``.}

\item{K}{Number of components modelled.}
}
\description{
Internal function that checks the initial labels are the correct
type.
}
