% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTypes.R
\name{checkTypes}
\alias{checkTypes}
\title{Check types}
\usage{
checkTypes(types)
}
\arguments{
\item{types}{Density types passed to ``callMDI``. Should be a vector of
strings. Viable options are ``'MVN'``, ``'TAGM'`` and ``'C'``.}
}
\description{
Internal function that checks the types passed to ``callMDI`` 
are the correct format.
}
