% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePSMsAcrossChains.R
\name{comparePSMsAcrossChains}
\alias{comparePSMsAcrossChains}
\title{Compare Posterior Similarity Matrices Across Chains}
\usage{
comparePSMsAcrossChains(
  mcmc,
  matrix_setting_order = 1,
  use_common_ordering = TRUE,
  ignore_checks = TRUE
)
}
\arguments{
\item{matrix_setting_order}{The matrix that defines the ordering of all
others, if a common ordering is used. Defaults to `1`.}

\item{use_common_ordering}{Logical indicating if a common ordering should be
used. If false all matrices are ordered individually, defaults to `TRUE`.}

\item{ignore_checks}{Logical indicating if checks for matrix symmetry and 
a commmon size should be ignored. Defaults to TRUE as these checks seem
to break when they should not.}

\item{mccm}{Output of ``processMCMCChains`` with the argument 
``construct_psm = TRUE``.}
}
\value{
A long data.frame containing columns `x` (the x-axis position of the
entry for geom_tile()), `y` (the y-axis position of the entry for
geom_tile()), `Entry` (value in similarity  matrix), `Chain` (assumes
chains are ordered from one to the number of chains present) and `View`.
}
\description{
Takes the PSMs from the output of ``processMCMCChains`` and
converts them to a data.frame prepared for ``ggplot2``.
}
\examples{

N <- 100
K <- 4
P <- 10

X <- generateSimulationDataset(K, N, P)
Y <- generateSimulationDataset(K, N, P)
Z <- generateSimulationDataset(K, N, P)

annotatedHeatmap(X$data, X$cluster_IDs, cluster_rows = FALSE)
annotatedHeatmap(Y$data, Y$cluster_IDs, cluster_rows = FALSE)
annotatedHeatmap(Z$data, Z$cluster_IDs, cluster_rows = FALSE)

row.names(Z$data) <- row.names(Y$data) <- row.names(X$data)

data_modelled <- list(X$data, Y$data, Z$data)

n_chains <- 4
R <- 1000
thin <- 25
types <- c("G", "G", "G")
K <- c(10, 10, 10)
mcmc <- runMCMCChains(data_modelled, n_chains, R, thin, types, K = K)

burn <- 250
mcmc <- processMCMCChains(mcmc, burn, construct_psm = TRUE)
psm_df <- comparePSMsAcrossChains(mcmc)

psm_df |> 
  ggplot(aes(x = x, y = y, fill = Entry)) +
  geom_tile() +
  facet_grid(View ~ Chain, labeller = label_both) +
  scale_fill_gradient(low = "#FFFFFF", high = "#146EB4") +
  labs(x = "Item", y = "Item", fill = "Coclustering\nproportion") +
  theme(
    axis.text = element_blank(),
    axis.ticks = element_blank(),
    panel.grid = element_blank(),
    axis.title.y = element_text(size = 10.5),
    axis.title.x = element_text(size = 10.5),
    plot.title = element_text(size = 18, face = "bold"),
    plot.subtitle = element_text(size = 14),
    strip.text.x = element_text(size = 10.5),
    legend.text = element_text(size = 10.5)
  )

}
