% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createSimilarityMat}
\alias{createSimilarityMat}
\title{Create Similarity Matrix}
\usage{
createSimilarityMat(allocations)
}
\arguments{
\item{allocations}{Matrix of sampled partitions. Columns correspond to 
items/samples being clustered, each row is a sampled partition.//'}
}
\value{
A symmetric n x n matrix (for n rows in cluster record) describing 
the fraction of iterations for which each pairwise combination of points are
assigned the same label.
}
\description{
Constructs a similarity matrix of the pairwise coclustering 
rate.
}
