% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOrder.R
\name{findOrder}
\alias{findOrder}
\title{Prepare data for ggplot heatmap}
\usage{
findOrder(X, dist = "euclidean", linkage = "complete")
}
\arguments{
\item{X}{Matrix.}

\item{dist}{Distance used when defining a distance matrix of X.}

\item{linkage}{Type of linkage used in the hierarchical clustering of X.}
}
\value{
The order of the rows of X based on a hierarchical clustering.
}
\description{
Find the ordering of the rows of X based on a hierarchical 
clustering of X for linkage ``linkage`` and distance ``dist``.
}
