% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gammaLogLikelihood}
\alias{gammaLogLikelihood}
\title{Gamma log-likelihood}
\usage{
gammaLogLikelihood(x, shape, rate)
}
\arguments{
\item{x}{- double; the value to calculate the likelihood of.}

\item{shape}{- double; the shape of the Gamma distribution.}

\item{rate}{- double; the rate of the Gamma distribution}
}
\value{
the log-likelihood of x in a Gamma with parameters shape 
and rate.
}
\description{
Used in calculating model probability in Metropolis-Hastings 
algorithm when proposals are from the Gamma distribution.
}
