% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gaussianLogLikelihood}
\alias{gaussianLogLikelihood}
\title{Multivariate normal log-likelihood for diagonal covariance matrix}
\usage{
gaussianLogLikelihood(x, mu, std_dev)
}
\arguments{
\item{x}{- vec; the sample to calculate the log likelihood of.}

\item{mu}{- vec; the mean parameter of the Gaussian distribution.}

\item{sigma_2}{- vec; the standard deviation of the Gaussian distribution.}
}
\value{
the normalised log-likelihood of x in a Gaussian distribution with 
parameters mu, sigma_2.
}
\description{
The log-likelihood function for a vector in a Gaussian density 
with a diagonal covariance matrix
}
