% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInitialLabels.R
\name{generateInitialLabels}
\alias{generateInitialLabels}
\title{Generate initial labels}
\usage{
generateInitialLabels(labels, fixed, K, alpha, labels_as_intended = FALSE)
}
\arguments{
\item{labels}{$N x V$ matrix of initial labels. The actual values passed only
matter for semi-supervised views (i.e. the views for which some labels are 
observed).}

\item{fixed}{$N x V$ matrix indicating which labels are observed and hence 
fixed. If the vth column has no 1's then this view is unsupervised.}

\item{K}{The number of components modelled in each  view.}

\item{alpha}{The concentration parameter (vector).}
}
\value{
An N vector of labels.
}
\description{
For V views, generate initial labels allowing for both 
semi-supervised and unsupervised views.
}
