% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateInitialUnsupervisedLabels.R
\name{generateInitialUnsupervisedLabels}
\alias{generateInitialUnsupervisedLabels}
\title{Generate initial unsupervised labels}
\usage{
generateInitialUnsupervisedLabels(N, alpha, K)
}
\arguments{
\item{N}{The number of items modelled in the dataset.}

\item{alpha}{The concentration parameter.}

\item{K}{The number of components modelled.}
}
\value{
An N vector of labels.
}
\description{
For a dataset with no observed labels, generate a vector of 
initial labels based on the stick-breaking prior for component weights.
}
