% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLikelihood.R
\name{getLikelihood}
\alias{getLikelihood}
\title{Get likelihood}
\usage{
getLikelihood(mcmc_output)
}
\arguments{
\item{mcmc_output}{The output from the mixture model.}
}
\value{
A data.frame containing the model log-likelihood and the associated
iteration.
}
\description{
Extracts the model fit score from the mixture model output.
}
