% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invGammaLogLikelihood}
\alias{invGammaLogLikelihood}
\title{Inverse gamma log-likelihood}
\usage{
invGammaLogLikelihood(x, shape, scale)
}
\arguments{
\item{x}{- double; the value to calculate the likelihood of.}

\item{shape}{- double; the shape of the inverse-Gamma distribution.}

\item{scale}{- double; the scale of the inverse-Gamma distribution}
}
\value{
the unnormalised log-likelihood of x in a inverse-Gamma with parameters 
shape and scale.
}
\description{
Used in calculating model probability in Metropolis-Hastings 
algorithm when proposals are from the inverse-Gamma distribution.
}
