% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invWishartLogLikelihood}
\alias{invWishartLogLikelihood}
\title{Inverse-Wishart log-likelihood}
\usage{
invWishartLogLikelihood(X, Psi, nu, P)
}
\arguments{
\item{X}{- matrix; the matrix to calculate the likelihood of.}

\item{Psi}{- matrix; the scale of the inverse-Wishart distribution.}

\item{nu}{- double; the degrees of freedom for the inverse-Wishart distribution.}

\item{P}{- unsigned integer; the dimension of X.}
}
\value{
the unnormalised log-likelihood of X in a inverse-Wishart with parameters Psi 
and nu.
}
\description{
Used in calculating model probability in Metropolis-Hastings 
algorithm when proposals are from the Wishart distribution.
}
