% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvtLogLikelihood}
\alias{mvtLogLikelihood}
\title{Multivariate t log-likelihood}
\usage{
mvtLogLikelihood(x, mu, Sigma, nu)
}
\arguments{
\item{x}{- vector; the sample to calculate the log likelihood of.}

\item{mu}{- vector; the mean parameter of the MVT distribution.}

\item{Sigma}{- matrix; the scale matrix of the MVT distribution.}

\item{nu}{- double; the degrees of freedom for the MVT distribution.}
}
\value{
the normalised log-likelihood of x in a MVT distribution with 
parameters mu, Sigma and nu.
}
\description{
The log-likelihood function for a point in the multivariate t 
(MVT) distribution.
}
