% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleMean}
\alias{sampleMean}
\title{Sample mean}
\usage{
sampleMean(X)
}
\arguments{
\item{X}{Matrix}
}
\value{
Vector of the column means of X.
}
\description{
calculate the sample mean of a matrix X.
}
