% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleStickBreakingPrior.R
\name{sampleStickBreakingPrior}
\alias{sampleStickBreakingPrior}
\title{Sample stick breaking prior}
\usage{
sampleStickBreakingPrior(alpha, K)
}
\arguments{
\item{alpha}{The concentration parameter.}

\item{K}{The number of weights to generate.}
}
\value{
A vector of weights.
}
\description{
Draw weights from the stick-breaking prior.
}
\examples{
weights <- stickBreakingPrior(1, 50)
}
