% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupOutlierComponents.R
\name{setupOutlierComponents}
\alias{setupOutlierComponents}
\title{Setup outlier components}
\usage{
setupOutlierComponents(types)
}
\arguments{
\item{types}{Density types passed to ``callMDI``.}
}
\value{
Vector of  numbers to pass to the C++ MDI function indicating what
outlier component types should be used. 0, the default, means no outliers are
modelled, 1 uses a global MVT distribution to model outliers a la Crook et al.
}
\description{
Internal function that translates the types passed to ``callMDI`` 
into the outlier component type vector.
}
