% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{squaredExponentialFunction}
\alias{squaredExponentialFunction}
\title{Squared exponential function}
\usage{
squaredExponentialFunction(amplitude, length, i, j)
}
\arguments{
\item{amplitude}{The amplitude parameter (double)}

\item{length}{The length parameter (double)}

\item{i}{Time point (unsigned integer)}

\item{j}{Time point (unsigned integer)}
}
\value{
Boolean indicating acceptance.
}
\description{
The squared exponential function as used in a covariance kernel.
}
