% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stickBreakingPrior.R
\name{stickBreakingPrior}
\alias{stickBreakingPrior}
\title{Stick breaking prior}
\usage{
stickBreakingPrior(alpha, K)
}
\arguments{
\item{alpha}{The concentration parameter.}

\item{K}{The number of weights to generate.}
}
\value{
A vector of weights.
}
\description{
Draw weights from the stick-breaking prior.
}
\examples{
weights <- stickBreakingPrior(1, 50)
}
