% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagmReDraft-package.R
\docType{package}
\name{tagmReDraft-package}
\alias{tagmReDraft-package}
\alias{tagmmdiReDraft}
\alias{tagmReDraft}
\title{Semi-supervised Multiple Dataset Integration}
\description{
Bayesian integrative clustering/classification. Model estimated
via a Gibbs sampler implemented in C++.
}
\author{
Stephen Coleman <stcolema@tcd.ie>, Paul D.W. Kirk, Chris Wallace,
Oliver M. Crook
}
\keyword{package}
