% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateTypes.R
\name{translateTypes}
\alias{translateTypes}
\title{Translate types}
\usage{
translateTypes(types)
}
\arguments{
\item{types}{Density types passed to ``callMDI``. Should be a vector of
strings. Viable options are ``'MVN'``, ``'TAGM'`` and ``'C'``.}
}
\value{
Vector of integers
}
\description{
Converts types from strings to integers passed to C++.
}
