% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wishartLogLikelihood}
\alias{wishartLogLikelihood}
\title{Wishart log-likelihood}
\usage{
wishartLogLikelihood(X, V, n, P)
}
\arguments{
\item{X}{- matrix; the matrix to calculate the likelihood of.}

\item{V}{- matrix; the scale of the Wishart distribution.}

\item{n}{- double; the degrees of freedom for the Wishart distribution.}

\item{P}{- unsigned integer; the dimension of X.}
}
\value{
the unnormalised log-likelihood of X in a Wishart with parameters V 
and n.
}
\description{
Used in calculating model probability in Metropolis-Hastings 
algorithm when proposals are from the Wishart distribution.
}
