% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllocProb.R
\name{calcAllocProb}
\alias{calcAllocProb}
\title{Calculate allocation probabilities}
\usage{
calcAllocProb(mcmc_samples, view, burn = 0, method = "median")
}
\arguments{
\item{mcmc_samples}{Output from ``callMDI``.}

\item{view}{Which view to calculate the allocation probabilities for.}

\item{burn}{The number of samples to discard.}

\item{method}{The point estimate to use. ``method = 'mean'`` or
``method = 'median'``. ``'median'`` is the default.}
}
\value{
An N x K matrix of class probabilities.
}
\description{
Calculate the empirical allocation probability for each class
based on the sampled allocation probabilities.
}
