% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFusionProbabiliy.R
\name{calcFusionProbabiliy}
\alias{calcFusionProbabiliy}
\title{Caluclate fusion probability}
\usage{
calcFusionProbabiliy(mcmc, v, w)
}
\arguments{
\item{mcmc}{Output from ``runMCMCChain``}

\item{v}{First view considered.}

\item{w}{Second view considered.}
}
\value{
A vector of probabilities
}
\description{
The mean of the number of MCMC samples for which each item has 
the same label in views v and w.
}
