% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFusionProbabiliyAllViews.R
\name{calcFusionProbabiliyAllViews}
\alias{calcFusionProbabiliyAllViews}
\title{Caluclate fusion probability all views}
\usage{
calcFusionProbabiliyAllViews(mcmc)
}
\arguments{
\item{mcmc}{Output from ``runMCMCChain``}
}
\value{
A list of vector of probabilities for each item fusing across each 
pair of views.
}
\description{
The mean of the number of MCMC samples for which each item has 
the same label in each pair of views.
}
