% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDataCorrectInput.R
\name{checkDataCorrectInput}
\alias{checkDataCorrectInput}
\title{Check data correct input}
\usage{
checkDataCorrectInput(X)
}
\arguments{
\item{X}{Data passed to ``callMDI``. Should be a list of matrices each with
N items held in rows.}
}
\value{
No return value, called for side effects.
}
\description{
Internal function that checks the data passed to ``callMDI`` is
the correct format.
}
\examples{
N <- 100
X <- matrix(c(rnorm(N, 0, 1), rnorm(N, 3, 1)), ncol = 2, byrow = TRUE)
Y <- matrix(c(rnorm(N, 0, 1), rnorm(N, 3, 1)), ncol = 2, byrow = TRUE)
data_modelled <- list(X, Y)
checkDataCorrectInput(data_modelled)
}
