% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFixedInput.R
\name{checkFixedInput}
\alias{checkFixedInput}
\title{Generate initial labels}
\usage{
checkFixedInput(fixed, N, V)
}
\arguments{
\item{fixed}{$N x V$ matrix indicating which labels are observed and hence 
fixed. If the vth column has no 1's then this view is unsupervised.}

\item{labels}{$N x V$ matrix of initial labels. The actual values passed only
matter for semi-supervised views (i.e. the views for which some labels are 
observed).}

\item{alpha}{The concentration parameter (vector).}

\item{K}{The number of components modelled in each  view.}
}
\description{
For V views, generate initial labels allowing for both 
semi-supervised and unsupervised views.
}
