% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareMSObject.R
\name{prepareMSObject}
\alias{prepareMSObject}
\title{Prepare MS Object}
\usage{
prepareMSObject(MS_object)
}
\arguments{
\item{MS_object}{A mass spectrometry experiment such as ``tan2009r1`` from 
``pRolocdata``.}
}
\value{
A list of ``X``, the fracitonation data from a LOPIT experiment, 
 ``fixed``, the matrix indicating which labels are observed,  
 ``initial_labels``, the matrix of the initial labels to be input into 
 ``runMCMCChains`` (note that ``"unknown"`` organelles are represented 
 arbitrarily with a ``1`` as these will be sampled again within the wrapper 
 of ``callMDI``) and ``class_key`` which maps the numeric representation of 
 organelles back to the original naming.
 @importFrom Biobase exprs fData
}
\description{
Prepares a mass spectrometry experiment (as stored in 
the Bioconductor package ``pRolocdata``) for modelling, extracting the 
numerical data, the classes and the indicator matrix for which labels are 
observed.
}
