% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSimilarityMatricesForGGplot.R
\name{prepSimilarityMatricesForGGplot}
\alias{prepSimilarityMatricesForGGplot}
\title{Prepare similarity matrices for ggplot}
\usage{
prepSimilarityMatricesForGGplot(
  similarity_matrices,
  matrix_setting_order = 1,
  use_common_ordering = TRUE,
  ignore_checks = FALSE
)
}
\arguments{
\item{similarity_matrices}{List of similarity matrices}

\item{matrix_setting_order}{The matrix that defines the ordering of all
others, if a common ordering is used. Defaults to `1`.}

\item{use_common_ordering}{Logical indicating if a common ordering should be
used. If false all matrices are ordered individually, defaults to `TRUE`.}
}
\value{
A long data.frame containing columns `x` (the x-axis position of the
entry for geom_tile()), `y` (the y-axis position of the entry for
geom_tile()), `Entry` (value in similarity  matrix) and `Chain` (assumes
chains are ordered from one to the number of chains present).
}
\description{
Given a list of symmatric similarity matrices, converts them
to the correct formatting for ggplot2, optionally giving all a common
ordering.
}
\examples{
plot_df <- prepSimilarityMatricesForGGplot(my_psms)
plot_df |>
  ggplot(aes(x = x, y = y, fill = Entry)) +
  geom_tile() +
  facet_wrap(~Chain) +
  scale_fill_gradient(low = "#FFFFFF", high = "#146EB4")
}
