% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMCMCChain.R
\name{processMCMCChain}
\alias{processMCMCChain}
\title{Process MCMC chain}
\usage{
processMCMCChain(mcmc_output, burn, point_estimate_method = "median")
}
\arguments{
\item{mcmc_output}{Output from ``batchSemiSupervisedMixtureModel``}

\item{burn}{The number of MCMC samples to drop as part of a burn in.}

\item{point_estimate_method}{Summary statistic used to define the point 
estimate. Must be ``'mean'`` or ``'median'``. ``'median'`` is the default.}
}
\value{
A named list similar to the output of 
``batchSemiSupervisedMixtureModel`` with some additional entries:
 * ``allocation_probability``: $(N x K)$ matrix. The point estimate of 
 the allocation probabilities for each data point to each class.
 
 * ``prob``: $N$ vector. The point estimate of the probability of being 
 allocated to the class with the highest probability.
 
 * ``pred``: $N$ vector. The predicted class for each sample.
}
\description{
Applies a burn in to and finds a point estimate for the output
of ``batchSemiSupervisedMixtureModel``.
}
